requirements.txt
Flask==2.0.1
requests==2.26.0
web3==5.24.0
py-solc-x==1.1.1
tabulate==0.8.9
argparse==1.4.0
python-dotenv==0.19.1 